sync: hosted cloud (#1540)
Submit a pull request for this project.

<!-- If you have an Issue that related to this Pull Request, you can
copy this Issue's description -->

# Why? 
<!-- 
> Related to which issue?
> Why we need this pull request?
> What is the user story for this pull request? 
-->


# What?
<!-- 
> Can you describe this feature in detail?
> Who can benefit from it? 
-->


# How?
<!-- 
> Do you have a simple description of how this pull request is
implemented?
-->

---------

+++ backend-server/application/src/main/java/com/apitable/shared/util/page/PageHelper.java
@@ -125,11 +125,11 @@ public class PageHelper {
     }
 
     public static <T> PageInfo<T> build(IPage<T> page) {
-        return new PageInfo<>((int) page.getCurrent(), (int) page.getSize(), (int) page.getTotal(),
+        return new PageInfo<>(page.getCurrent(), page.getSize(), page.getTotal(),
             page.getRecords());
     }
 
-    public static <T> PageInfo<T> build(int pageNum, int pageSize, int total, List<T> records) {
+    public static <T> PageInfo<T> build(long pageNum, long pageSize, long total, List<T> records) {
         return new PageInfo<>(pageNum, pageSize, total, records);
     }
 }
